Add evaluation.rb. created by @honbin
