+++ app/models/post_evaluation.rb
@@ -17,18 +17,20 @@ class PostEvaluation < ActiveRecord::Base
   scope :good, where(kind: '1')
   scope :bad, where(kind: '2')
 
-  def kind_list
+  include Evaluation
+
+  #def kind_list
     # "1": good
     # "2": bad
-    [ "1", "2" ]
-  end
+  #  [ "1", "2" ]
+  #end
 
-private
-  def include_by_kind?
-    kind_list.include? kind
-  end
-
-  def kind_valid?
-    errors.add(:kind, 'kindが不正') unless include_by_kind?
-  end
+#private
+#  def include_by_kind?
+#    kind_list.include? kind
+#  end
+#
+#  def kind_valid?
+#    errors.add(:kind, 'kindが不正') unless include_by_kind?
+#  end
 end
